requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
scikit-learn
yfinance
alpha_vantage
plotly
joblib
matplotlib
pytest
python-dotenv
torch
shap
tsfresh
statsmodels
pandas-profiling
xlrd
openpyxl